requirements.txt
google-auth
google-auth-oauthlib
google-cloud-storage
google-cloud-firestore
google-api-python-client
